+++ cmd/formulaDetails.go
@@ -17,8 +17,6 @@ package cmd
 import (
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
-	"net/http"
 	"os"
 
 	"github.com/ghchinoy/cectl/ce"
@@ -57,29 +55,11 @@ var formulaDetailsCmd = &cobra.Command{
 		*/
 		auth := fmt.Sprintf("User %s, Organization %s", user, org)
 
-		bodybytes, statuscode, err := getFormulaDetails(args[0], fmt.Sprintf("%s", base), auth)
+		bodybytes, statuscode, err := ce.FormulaDetailsAsBytes(args[0], fmt.Sprintf("%s", base), auth)
 		if err != nil {
 			fmt.Println("unable to retrieve formula", err.Error())
 			os.Exit(1)
 		}
-		/*
-			client := &http.Client{}
-			req, err := http.NewRequest("GET", url, nil)
-			if err != nil {
-				fmt.Println("Can't construct request", err.Error())
-				os.Exit(1)
-			}
-			req.Header.Add("Authorization", auth)
-			req.Header.Add("Accept", "application/json")
-			req.Header.Add("Content-Type", "application/json")
-			resp, err := client.Do(req)
-			if err != nil {
-				fmt.Println("Cannot process response", err.Error())
-				os.Exit(1)
-			}
-			bodybytes, err := ioutil.ReadAll(resp.Body)
-			defer resp.Body.Close()
-		*/
 
 		if outputJSON {
 			fmt.Printf("%s\n", bodybytes)
@@ -112,36 +92,6 @@ var formulaDetailsCmd = &cobra.Command{
 		}
 
 	},
-}
-
-func getFormulaDetails(formulaID, base, auth string) ([]byte, int, error) {
-
-	var bodybytes []byte
-
-	url := fmt.Sprintf("%s%s",
-		base,
-		fmt.Sprintf(ce.FormulaURIFormat, formulaID),
-	)
-
-	client := &http.Client{}
-	req, err := http.NewRequest("GET", url, nil)
-	if err != nil {
-		fmt.Println("Can't construct request", err.Error())
-		os.Exit(1)
-	}
-	req.Header.Add("Authorization", auth)
-	req.Header.Add("Accept", "application/json")
-	req.Header.Add("Content-Type", "application/json")
-	resp, err := client.Do(req)
-	if err != nil {
-		return bodybytes, resp.StatusCode, err
-
-	}
-	bodybytes, err = ioutil.ReadAll(resp.Body)
-	defer resp.Body.Close()
-
-	return bodybytes, resp.StatusCode, nil
-
 }
 
 func init() {